+++ std/jwt/test.ts
@@ -15,36 +15,9 @@ import {
 } from "./validate.ts";
 
 import { assertEquals, assertThrows } from "../testing/asserts.ts";
-import { convertHexToBase64url, isExpired } from "./_util.ts";
+import { convertHexToBase64url } from "./_util.ts";
 
 const key = "your-secret";
-
-Deno.test("[jwt] isExpired", function (): void {
-  // A specific date:
-  const t1 = setExpiration(new Date("2020-01-01"));
-  const t2 = setExpiration(new Date("2099-01-01"));
-  // Ten seconds from now:
-  const t3 = setExpiration(10);
-  // One hour from now:
-  const t4 = setExpiration(60 * 60);
-  //  1 second from now:
-  const t5 = setExpiration(1);
-  //  1 second earlier:
-  const t6 = setExpiration(-1);
-  assertEquals(isExpired(t1), true);
-  assertEquals(isExpired(t2), false);
-  assertEquals(10, t3 - Math.round(Date.now() / 1000));
-  assertEquals(isExpired(t4), false);
-  assertEquals(isExpired(t5), false);
-  assertEquals(isExpired(t6), true);
-  // add leeway:
-  assertEquals(isExpired(t6, 1500), false);
-  assertEquals(setExpiration(10), setExpiration(new Date(Date.now() + 10000)));
-});
-
-Deno.test("[jwt] setExpiration", function (): void {
-  assertEquals(setExpiration(10), setExpiration(new Date(Date.now() + 10000)));
-});
 
 Deno.test("[jwt] makeSignature", async function (): Promise<void> {
   // https://www.freeformatter.com/hmac-generator.html
@@ -99,13 +72,13 @@ Deno.test("[jwt] parse", async function (): Promise<void> {
   );
   assertThrows(() => {
     parse(".aaa.bbb");
-  }, SyntaxError);
+  }, Error);
   assertThrows((): void => {
     parse("a..aa.bbb");
   }, TypeError);
   assertThrows((): void => {
     parse("aaa.bbb.ccc.");
-  }, SyntaxError);
+  }, Error);
   const jwt =
     "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c";
   const header:Header = {
